requirements.txt
Flask
Flask-Login
Flask-WTF
Flask-SQLAlchemy
email-validator
psycopg2-binary
python-dotenv
